+++ src/linearmixedmodel.jl
@@ -377,7 +377,7 @@ end
 StatsBase.fitted(m::LinearMixedModel{T}) where {T} = fitted!(Vector{T}(undef, nobs(m)), m)
 
 """
-    fixef!(v::Vector{T}, m::LinearMixedModel{T})
+    fixef!(v::Vector{T}, m::MixedModel{T})
 
 Overwrite `v` with the pivoted fixed-effects coefficients of model `m`
 
@@ -494,15 +494,6 @@ function Base.getproperty(m::LinearMixedModel{T}, s::Symbol) where {T}
         getfield(m, s)
     end
 end
-
-"""
-    issingular(m::LinearMixedModel, θ=m.θ)
-
-Test whether the model `m` is singular if the parameter vector is `θ`.
-
-Equality comparisons are used b/c small non-negative θ values are replaced by 0 in `fit!`.
-"""
-issingular(m::LinearMixedModel, θ=m.θ) = any(lowerbd(m) .== θ)
 
 function StatsBase.leverage(m::LinearMixedModel{T}) where {T}
     # This can be done more efficiently but reusing existing tools is easier.