+++ src/simulate.jl
@@ -1,6 +1,6 @@
 """
-    simulate!(rng::AbstractRNG, m::LinearMixedModel{T}; β=m.β, σ=m.σ, θ=T[])
-    simulate!(m::LinearMixedModel; β=m.β, σ=m.σ, θ=m.θ)
+    simulate!(rng::AbstractRNG, m::MixedModel{T}; β=m.β, σ=m.σ, θ=T[])
+    simulate!(m::MixedModel; β=m.β, σ=m.σ, θ=m.θ)
 
 Overwrite the response (i.e. `m.trms[end]`) with a simulated response vector from model `m`.
 """
@@ -33,11 +33,111 @@ function simulate!(
         unscaledre!(rng, y, trm)
     end
                     # scale by σ and add fixed-effects contribution
-    BLAS.gemv!('N', one(T), m.X, β, σ, y)
+    mul!(y, m.X, β, one(T), σ)
+
     m
 end
 
-function simulate!(m::LinearMixedModel{T}; β = coef(m), σ = m.σ, θ = T[]) where {T}
+function simulate!(
+    rng::AbstractRNG,
+    m::GeneralizedLinearMixedModel{T};
+    β = coef(m),
+    σ = m.σ,
+    θ = T[],
+) where {T}
+    length(β) == length(fixef(m)) ||
+        length(β) == length(coef(m)) ||
+            throw(ArgumentError("You must specify all (non-singular) βs"))
+
+    dispersion_parameter(m) || ismissing(σ) ||
+        throw(ArgumentError("You must not specify a dispersion parameter for model families without a dispersion parameter"))
+
+    β = convert(Vector{T},β)
+    if σ !== missing
+        σ = T(σ)
+    end
+    θ = convert(Vector{T},θ)
+
+    d = m.resp.d
+
+    if length(β) ≠ length(coef(m))
+        padding = length(coef(m)) - length(β)
+        for ii in 1:padding
+            push!(β, -0.0)
+        end
+    end
+
+    fast = (length(m.θ) == length(m.optsum.final))
+    setpar! = fast ? setθ! : setβθ!
+    params = fast ? θ : vcat(β, θ)
+    setpar!(m, params)
+
+    lm = m.LMM
+
+    # note that these m.resp.y and m.LMM.y will later be sychronized in (re)fit!()
+    # but for now we use them as distinct scratch buffers to avoid allocations
+
+    # the noise term is actually in the GLM and not the LMM part so no noise
+    # at the LMM level
+    η = fill!(m.LMM.y, zero(T))
+    y = m.resp.y
+
+    # assemble the linear predictor
+
+    # add the unscaled random effects
+    # note that unit scaling may not be correct for
+    # families with a dispersion parameter
+    @inbounds for trm in m.reterms
+        unscaledre!(rng, η, trm)
+    end
+
+    # add fixed-effects contribution
+    # note that unit scaling may not be correct for
+    # families with a dispersion parameter
+    mul!(η, lm.X, β, one(T), one(T))
+
+    # from η to μ
+    GLM.updateμ!(m.resp, η)
+
+    # convert to the distribution / add in noise
+    @inbounds for (idx, val) in enumerate(m.resp.mu)
+        n = isempty(m.wt) ? 1 : m.wt[idx]
+        y[idx] = _rand(rng, d, val, σ, n)
+    end
+
+    m
+end
+
+"""
+    _rand(rng::AbstractRNG, d::Distribution, location, scale=missing, n=1)
+
+A convenience function taking a draw from a distrbution.
+
+Note that `d` is specified as an existing distribution, such as
+from the `GlmResp.d` field. This isn't vectorized nicely because
+for distributions where the scale/dispersion is dependent on the
+location (e.g. Bernoulli, Binomial, Poisson), it's not really
+possible to avoid creating multiple `Distribution` objects.
+
+Note that `n` is the `n` parameter for the Binomial distribution,
+*not* the number of draws from the RNG. It is then used to change the
+random draw (an integer in [0, n]) into a probability (a float in [0,1]).
+"""
+function _rand(rng::AbstractRNG, d::Distribution, location, scale=NaN, n=1)
+    if !ismissing(scale)
+        throw(ArgumentError("Families with a dispersion parameter not yet supported"))
+    end
+
+    if d isa Binomial
+        dist = Binomial(Int(n), location)
+    else
+        dist = typeof(d)(location)
+    end
+
+    rand(rng, dist) / n
+end
+
+function simulate!(m::MixedModel{T}; β = coef(m), σ = m.σ, θ = T[]) where {T}
     simulate!(Random.GLOBAL_RNG, m, β = β, σ = σ, θ = θ)
 end